minor code standards fixes

+++ procedures/day/delete.php
@@ -4,10 +4,10 @@ $returnData = array();
 
 $returnData['valid'] = 0;
 
-$guid = get_input("guid");
+$guid = (int) get_input("guid");
 
 if (!empty($guid) && ($eventDay = get_entity($guid))) {
-	if ($eventDay->getSubtype() == EventDay::SUBTYPE) 	{
+	if ($eventDay->getSubtype() == EventDay::SUBTYPE) {
 		if ($eventDay->delete()) {
 			$returnData['valid'] = 1;
 		}
